Resolve problem with deleting entities with many-to-many relation, update tests, update specification

+++ src/main/java/ru/task_manager/services/ProjectService.java
@@ -3,7 +3,6 @@ package ru.task_manager.services;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.task_manager.entities.ProjectEntity;
-import ru.task_manager.entities.TaskEntity;
 import ru.task_manager.entities.UserEntity;
 import ru.task_manager.exceptions.ProjectNotFoundException;
 import ru.task_manager.exceptions.UserNotFoundException;
@@ -11,18 +10,20 @@ import ru.task_manager.repositories.ProjectRepo;
 import ru.task_manager.repositories.TaskRepo;
 import ru.task_manager.repositories.UserRepo;
 
-import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
+
 
 @Service
 @RequiredArgsConstructor
 public class ProjectService {
     private final ProjectRepo projectRepo;
     private final UserRepo userRepo;
-    private final TaskRepo taskRepo;
+    private final EntityRelationService entityRelationService;
 
     public Long createProject(ProjectEntity projectEntity, List<Long> usersIdList) {
-        List<UserEntity> userEntityList = new ArrayList<>();
+        Set<UserEntity> userEntityList = new HashSet<>();
         for (Long userId : usersIdList) {
             userEntityList.add(userRepo.findById(userId)
                     .orElseThrow(() -> new UserNotFoundException(userId.toString())));
@@ -32,7 +33,7 @@ public class ProjectService {
     }
 
     public ProjectEntity updateProject(ProjectEntity projectEntity, List<Long> usersIdList) {
-        List<UserEntity> userEntityList = new ArrayList<>();
+        Set<UserEntity> userEntityList = new HashSet<>();
         for (Long userId : usersIdList) {
             userEntityList.add(userRepo.findById(userId)
                     .orElseThrow(() -> new UserNotFoundException(userId.toString())));
@@ -51,19 +52,9 @@ public class ProjectService {
     }
 
     public void delete(Long id) {
-        ProjectEntity projectEntity = getProjectById(id);
-        List<UserEntity> userEntities = projectEntity.getUsersEntity();
-        for(UserEntity userEntity: userEntities){
-            userEntity.deleteProjectFromList(projectEntity);
-            userRepo.save(userEntity);
-        }
-
-        List<TaskEntity> taskEntities = projectEntity.getTaskEntities();
-        for (TaskEntity taskEntity: taskEntities){
-            taskEntity.deleteProject();
-            taskRepo.save(taskEntity);
-        }
-
+        if (!projectRepo.existsById(id)){throw new ProjectNotFoundException(id.toString());}
+        entityRelationService.removeProjectFromUsers(id);
+        entityRelationService.removeProjectFromTasks(id);
         projectRepo.deleteById(id);
     }
